requirements.txt
Babel==2.3.4
blinker==1.4
Flask==0.11.1
Flask-Babel==0.11.1
Flask-Cache==0.13.1
Flask-Compress==1.3.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Moment==0.2.1
Flask-Principal==0.4.0
Flask-Script==0.6.7
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
Flask-OpenID==1.2.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib==1.6.5
pytz==2016.7
six==1.10.0
SQLAlchemy==1.0.15
SQLAlchemy-migrate==0.10.0
Werkzeug==0.11.11
WTForms==2.1
oauth2client==3.0.0
requests==2.11.1
httplib2==0.9.2


